Struct persistentcache::storage::redis::RedisStorage

pub struct RedisStorage { /* fields omitted */ }

RedisStorage struct holds a redis::Connection variable.

Methods

impl RedisStorage

pub fn new(host: &str) -> Result<Self, Error>

Connects to the Redis server listening at host and constructs a new RedisStorage struct.

This will fail in case there is no redis server running.

Examples

use persistentcache::storage::redis::RedisStorage;

let s = RedisStorage::new("redis://127.0.0.1").unwrap();